[relalg] Fix the most mysterious bug I've ever encountered in Vim.
And this is still a mystery to me.

After introducing non-transactional operations in librelalg, I have
started to get some seemingly random errors when building some
Colortemplate templates. Specifically, I have encountered the
weird situation detailed below when compiling all Solarized 8 templates
at once (with *BuildAll!*). No other templates give me errors, and
building each Solarized 8 template in isolation (with *Build*) never
fails either. Besides, from time to time even *BuildAll!* works! Crazy.

Here are the details of my investigation. Building Solarized 8 templates
with Colortemplate's *BuildAll!* has started to fail (most of the time, but
not always!) with this error:

```
[Colortemplate] Build failed: Uniqueness failed for {NormalizedName: 'vimcommentstring', HiGroup: 'vimCommentString', DiscrName: 'extra'}. Duplicate key: {HiGroup: 'vimCommentString'} already exists in Highlight Group. in included file "/Users/nicola/.vim/pack/colors/start/solarized8/templates/_extra.colortemplate", byte 854) (line 31, byte 855).
```

Each Solarized 8 template contains exactly two definitions for `VimCommentString`:

```
vimCommentString -> omit ; Skip default definition
vimCommentString  /gui/256/16 +extra 1     violet            none
```

What I have found is that when `Database.InsertHighlightGroup_()` (in
`colortemplate/import/colortemplate/colorscheme.vim`) performs a lookup
for `vimcommentstring`, such lookup returns `KEY_NOT_FOUND` for the
second definition, although the tuple exists because it was inserted when
the first definition was encountered.

Since the lookup fails, the method tries an `Insert()` into `HighlightGroup`,
which obviously causes a duplicate key error. Strangely enough, if the insert
is wrapped into a `Transaction()`, the error disappears. My guess is that using
transactional inserts avoids the strange phenomenon described below, but
I have absolutely no idea why.

Anyway, why does the lookup fail to find an existing tuple? Are the indexes
perhaps corrupted? Well, it turns out that they are. For some reason—totally
obscure to me—if I print the index on `NormalizedName`, I see that there is
an entry whose key is `['vimcommentstring', 'vimcommentstring']`. WTH?!
Why does this key have two items instead of one, given that the index is on
a single attribute (`NormalizedName`)?

You know what? When a tuple is inserted into the index (`Index.Add()` method
of librelalg), the key values are retrieved with:

```
 var keyValues = string(Values(t, this.key))
```

Well, what I have found is that, even if `this.key` is a list with one item
(`NormalizedName`), `Values()` returns two items. That is, the `for` loop
inside `Values()` is executed twice even if the `attrs` paramter is a list with one
element. This is totally nuts.

By chance, I've discovered that by accessing the `attrs` parameter inside
`Values()` before the `for` loop seems to fix the issue. So, that's the magical fix
implemented by this commit.

Maybe one day I will know why.

Signed-off-by: Lifepillar <lifepillar@lifepillar.me>